criacao nos serviços e implementações

+++ src/main/java/com/starking/vendas/services/dtos/ItemPedidoDTO.java
@@ -10,8 +10,8 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ItemPedidoDTO {
-	
-	private Integer produto;
+
+	private Long produto;
 	private Integer quantidade;
 
 }
